+++ graph-commons/src/main/scala/ch/datascience/http/server/security/Authenticator.scala
@@ -16,22 +16,11 @@
  * limitations under the License.
  */
 
-package ch.datascience.http.server
+package ch.datascience.http.server.security
 
 import ch.datascience.http.client.AccessToken
-import ch.datascience.http.client.AccessToken.{OAuthAccessToken, PersonalAccessToken}
-import org.http4s.AuthScheme.Bearer
-import org.http4s.Credentials.Token
-import org.http4s.Header
-import org.http4s.headers.Authorization
+import ch.datascience.http.server.security.model.AuthUser
 
-package object security {
-
-  implicit class AccessTokenOps(accessToken: AccessToken) {
-
-    lazy val toHeader: Header = accessToken match {
-      case PersonalAccessToken(token) => Header("PRIVATE-TOKEN", token)
-      case OAuthAccessToken(token)    => Authorization(Token(Bearer, token))
-    }
-  }
+trait Authenticator[Interpretation[_]] {
+  def authenticate(accessToken: AccessToken): Interpretation[Either[EndpointSecurityException, AuthUser]]
 }
