+++ knowledge-graph/src/main/scala/ch/datascience/knowledgegraph/MicroserviceRoutes.scala
@@ -23,6 +23,7 @@ import cats.effect.{Clock, ConcurrentEffect, ContextShift, IO, Resource, Timer}
 import cats.syntax.all._
 import ch.datascience.config.GitLab
 import ch.datascience.control.{RateLimit, Throttler}
+import ch.datascience.graph.http.server.security.{GitLabAuthenticator, ProjectAuthorizer}
 import ch.datascience.graph.model
 import ch.datascience.http.rest.SortBy.Direction
 import ch.datascience.http.rest.paging.PagingRequest
@@ -30,7 +31,7 @@ import ch.datascience.http.rest.paging.PagingRequest.Decoders._
 import ch.datascience.http.rest.paging.model.{Page, PerPage}
 import ch.datascience.http.server.QueryParameterTools._
 import ch.datascience.http.server.security.model.AuthUser
-import ch.datascience.http.server.security.{Authentication, ProjectAuthorizer}
+import ch.datascience.http.server.security.Authentication
 import ch.datascience.knowledgegraph.datasets.rest.DatasetsSearchEndpoint.Query.Phrase
 import ch.datascience.knowledgegraph.datasets.rest._
 import ch.datascience.knowledgegraph.graphql.{IOQueryEndpoint, QueryEndpoint}
@@ -147,7 +148,8 @@ private object MicroserviceRoutes {
       projectDatasetsEndpoint <- IOProjectDatasetsEndpoint(sparqlTimeRecorder)
       datasetEndpoint         <- IODatasetEndpoint(sparqlTimeRecorder)
       datasetsSearchEndpoint  <- IODatasetsSearchEndpoint(sparqlTimeRecorder)
-      authMiddleware          <- Authentication.middleware(gitLabThrottler, logger)
+      authenticator           <- GitLabAuthenticator(gitLabThrottler, logger)
+      authMiddleware          <- Authentication.middleware(authenticator)
       projectAuthorizer       <- ProjectAuthorizer(sparqlTimeRecorder, logger = logger)
       routesMetrics = new RoutesMetrics[IO](metricsRegistry)
     } yield new MicroserviceRoutes(queryEndpoint,
